Now works for full game, with 2 bots and 1 command line player
play until everyone except 1 player is out of chips ( so a full game).
right now with 2 naive bots (1/6 chance to fold, 2/6 chance to call, 2/6
chance to raise by *2, and 1/6 to go all in). and a command line player
(we will change command line player to our bot once that is doing
something)

output can be pretty hard to read, but we can fiddle with that later